more progress on test filter; stopping here to go fix login system
I will need a better authentication management system to make the test filter and
AJAX updating of the drop downs to work. Basically I need to ensure that I'm logged
in when I display a page so I don't have to worry about authentication and reloading
stuff.

Also modified .gitignore to exclude .dll and .pdb, as excluding the directories
did not seem to work. I removed the .dll and .pdb from the cache as well.